+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.data.mongodb.uri=mongodb://localhost:32768/twitter_db
 
-TWITTER_CONSUMER_KEY=IwBJTCqVCY4IgI7Pws03O3aGy
-TWITTER_CONSUMER_SECRET=tRdgtvZk3BtgQ4SKCY6e1Uz5P0yEm29xWaLGaszLri9UcUSpEm
-TWITTER_ACCESS_TOKEN=1175804596079476736-0oThCev4GvRbWJThs9EsDNKcG2jmML
-TWITTER_ACCESS_TOKEN_SECRET=vgGuUjvOzwYwbhx5aY3JJtfsG3e6exUHwnWyymVbL9HAG
+twitter.consumer-key=IwBJTCqVCY4IgI7Pws03O3aGy
+twitter.consumer-secret=tRdgtvZk3BtgQ4SKCY6e1Uz5P0yEm29xWaLGaszLri9UcUSpEm
+twitter.access-token=1175804596079476736-0oThCev4GvRbWJThs9EsDNKcG2jmML
+twitter.access-token-secret=vgGuUjvOzwYwbhx5aY3JJtfsG3e6exUHwnWyymVbL9HAG
